% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initProject.R
\name{initProject}
\alias{initProject}
\title{Init Project Addin}
\usage{
initProject()
}
\value{
the return of \code{\link[shiny]{runGadget}}
}
\description{
Initialize a structure for a new project
}
\examples{
\dontrun{
initProject()
}
}
